Refs 10837. add topic types to coverage ignore (with rename)
