+++ .github/workflows/test.yml
@@ -61,6 +61,7 @@ jobs:
           --exclude='.*docs/.*' \
           --exclude='.*test/.*' \
           --exclude='.*github/.*' \
+          --exclude='.*topic_types/.*' \
           --exclude-unreachable-branches
 
       - name: Upload coverage
