+++ src/cpp/topic_types/types.cxx
