+++ src/cpp/topic_types/types.h
