+++ test/TrafficInjector/MessageSerializer.h
@@ -19,7 +19,7 @@
 #ifndef _EPROSIMA_FASTDDS_STATISTICS_BACKEND_TEST_MESSAGE_SERIALIZER_HPP_
 #define _EPROSIMA_FASTDDS_STATISTICS_BACKEND_TEST_MESSAGE_SERIALIZER_HPP_
 
-#include <subscriber/types.h>
+#include <topic_types/types.h>
 
 #include <nlohmann-json/json.hpp>
